Bug: saving benchmarking scores to db and submission status updated not working after successful run trough  (#34)
* - fixed backend saving of benchmarking modules to db
-adapted submission model
-added unique constraint to benchmarking model to avoid duplicate entries

* - fixed detailview barchart

+++ backend/app/models/submission.py
@@ -12,7 +12,7 @@ class Submission(db.Model):
     submission_date = db.Column(db.DateTime, default=datetime.utcnow)
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     status = db.Column(db.Enum(SubmissionStatus), nullable=False)
-    score = db.Column(db.Integer, nullable=False)
+    score = db.Column(db.Float, nullable=False)
     is_public = db.Column(db.Boolean, nullable=False)
     created_at = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
